Customer registration services finished

+++ Models/Item.cs
@@ -13,9 +13,9 @@ namespace ZofyaApi.Models
 
         public string SKU { get; set; } = null!;
         public string Description { get; set; } = null!;
-        public double Discount { get; set; }
+        public decimal Discount { get; set; }
         public string Name { get; set; } = null!;
-        public double Price { get; set; }
+        public decimal Price { get; set; }
         public string Category { get; set; } = null!;
         public string Status { get; set; } = null!;
         public int Stock { get; set; }
